[wip] Load compressed sequences into database
This was tried, and ultimately a bit much overhead for `augur filter`.

The usage of sequence data in `augur filter` is trivial - all the logic
is done on the tabular metadata file which loads nicely into the
database. Sequence data is simply iterated via a generator, and anything
that has passed metadata and sequence index filters gets written out.
This means no sequence data needs to be kept in memory, eliminating the
need for loading all into a database, which comes with storage overhead:

Sequence data at large scale should be compressed. Uncompressed open
(GenBank) SARS-CoV-2 data as of today yielded a SQLite DB file of >80GB
before I terminated the process due to limited local storage.

The compression of all open (GenBank) SARS-CoV-2 sequences to date using
zlib took 4.5 hours on my local machine, compared to 7 minutes using the
current approach of reading and outputting on the fly.