Simplify command name parsing
Removes the `command_name` function to simplify the command name parsing
by directly replacing underscores with dashes. This allows us to create
the correct CLI command name for a command regardless of whether the
command is a module or a package and without the "unnecessary machinery."

Note this change was built on a suggestion by @tsibley in another PR¹,
but I'm incorporating here to support the reorganization of commands
with subcommands into packages. I chose to do the string replacement for
the command name rather than the command module so that we can directly
list module names and make the computer do the snake_case to kebab-case
conversion for us.

¹ https://github.com/nextstrain/augur/pull/997#issuecomment-1175522437